Moved grading logic to internal function in classController to prevent internal axios call
