+++ backend/.env.example
@@ -4,6 +4,5 @@ JWT_SECRET="YOUR_JWT_SECRET"
 GOOGLE_CLIENT_ID="YOUR_GOOGLE_CLIENT_ID"
 GOOGLE_CLIENT_SECRET="YOUR_GOOGLE_CLIENT_SECRET"
 FRONTEND_URL="YOUR_FRONTEND_URL"
-BACKEND_URL="YOUR_BACKEND_URL"
 GRADIA_API_KEY="YOUR_GRADIA_API_KEY"
 GRADIA_PYTHON_BACKEND_URL="YOUR_GRADIA_PYTHON_BACKEND_URL"
