--- backend/routes/gradingRoutes.js
@@ -1,8 +1,0 @@
-import express from "express";
-const router = express.Router();
-import { gradeSubmission } from "../controllers/gradingController.js";
-
-router.post("/grade-submission", gradeSubmission);
-
-export default router;
-
