+++ backend/server.js
@@ -6,7 +6,6 @@ import connectDb from './config/db.js';
 import authRoutes from "./routes/authRoutes.js";
 import classRoutes from './routes/classRoutes.js';
 import testRoutes from './routes/testRoutes.js';
-import gradingRoutes from './routes/gradingRoutes.js';
 import passport from "passport";
 import session from "express-session";
 import { setupGoogleAuth } from "./config/passport.js";
@@ -55,7 +54,6 @@ setupGoogleAuth("teacher");
 app.use("/api/auth", authRoutes);
 app.use("/api/classes", authMiddleware, classRoutes);
 app.use("/api/tests", authMiddleware, testRoutes);
-app.use("/api/grade", gradingRoutes);
 
 const PORT = process.env.PORT || 8000;
